requirements/requirements.txt
python ~>3.8
setuptools-rust
openai-whisper
git+https://github.com/openai/whisper.git
ffmpeg
tiktoken
tokenizers
torch
transformers
google-cloud-aiplatform
google-generativeai